# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ecommarce API
  version: 1.1.0
  description: |
    Versioned REST API serving both storefront and admin.
    Step 1: skeleton. Step 2: auth/rbac/audit/outbox. Step 3: catalog/inventory/search.
servers:
  - url: /api/v1
...
paths:
...
  /variants/product/{productId}:
    get:
      tags: [catalog]
      summary: List variants for a product
      security: []
      parameters:
        - { name: productId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Variant' }
    post:
      tags: [catalog]
      summary: Create variant for a product
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: productId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVariantDto' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Variant' }
  /variants/matrix:
    post:
      tags: [catalog]
      summary: Generate cartesian variant matrix (color x size etc.)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VariantMatrixRequest' }
      responses:
        '201':
          description: Created variants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Variant' }
  /variants/{id}:
    get:
      tags: [catalog]
      summary: Variant by id
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Variant' }
    patch:
      tags: [catalog]
      summary: Update variant
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateVariantDto' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Variant' }
    delete:
      tags: [catalog]
      summary: Delete variant
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Deleted }

  # ---------------------------------------------------------------------------
  # MEDIA (Step 3 stub, S3 in Step 12)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Variant:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        warehouseId: { type: string, nullable: true }
        sku: { type: string }
        barcode: { type: string, nullable: true }
        pricePoisha: { type: integer }
        compareAtPoisha: { type: integer, nullable: true }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        attributeValues:
          type: object
          additionalProperties: { type: string }
        isActive: { type: boolean }
    CreateVariantDto:
      type: object
      required: [sku, pricePoisha, attributeValues]
      properties:
        sku: { type: string }
        barcode: { type: string }
        pricePoisha: { type: integer }
        compareAtPoisha: { type: integer }
        stock: { type: integer }
        lowStockThreshold: { type: integer }
        attributeValues:
          type: object
          additionalProperties: { type: string }
        isActive: { type: boolean }
        warehouseId: { type: string }
    UpdateVariantDto: { $ref: '#/components/schemas/CreateVariantDto' }
    VariantMatrixRequest:
      type: object
      required: [productId, axes, basePricePoisha]
      properties:
        productId: { type: string }
        axes:
          type: array
          items:
            type: object
            required: [attributeSlug, values]
            properties:
              attributeSlug: { type: string }
              values:
                type: array
                items: { type: string }
        basePricePoisha: { type: integer }
        baseCompareAtPoisha: { type: integer }
        skuPrefix: { type: string }
        defaultStock: { type: integer }
...